requirements.txt
# PyTorch устанавливается отдельно:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128

transformers>=4.35.0
facenet-pytorch>=2.5.3
opencv-python>=4.8.0
librosa>=0.10.1
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pyyaml>=6.0
tqdm>=4.66.0
Pillow>=10.0.0
einops>=0.7.0
sounddevice>=0.4.6
